RNN and simulation refactored
